Resync on 17JUL, but no changes since last.
Tried to incorporate the changes that caldroid did in the past year.  Ran into issues when trying to do our custom fit of the calendar as it was causing the red box for the current day to disappear, but figured it out that it was the CellView padding that needed to be fixed (increased, as it was hiding the underlying background resource, which was the red box).  Also put in some other tests for getting a more proper window size in the CalendarActivity.